Move base RSS code to Core.
