+++ src/base/rss/rssmanager.h
