+++ src/gui/rss/rss.pri
@@ -3,31 +3,15 @@ INCLUDEPATH += $$PWD
 HEADERS +=    $$PWD/rss_imp.h \
               $$PWD/rsssettingsdlg.h \
               $$PWD/feedlistwidget.h \
-              $$PWD/rssmanager.h \
-              $$PWD/rssfeed.h \
-              $$PWD/rssfolder.h \
-              $$PWD/rssfile.h \
-              $$PWD/rssarticle.h \
               $$PWD/automatedrssdownloader.h \
-              $$PWD/rssdownloadrule.h \
-              $$PWD/rssdownloadrulelist.h \
               $$PWD/cookiesdlg.h \
-              $$PWD/rssparser.h \
               $$PWD/htmlbrowser.h
 
 SOURCES +=   $$PWD/rss_imp.cpp \
              $$PWD/rsssettingsdlg.cpp \
              $$PWD/feedlistwidget.cpp \
-             $$PWD/rssmanager.cpp \
-             $$PWD/rssfeed.cpp \
-             $$PWD/rssfolder.cpp \
-             $$PWD/rssarticle.cpp \
              $$PWD/automatedrssdownloader.cpp \
-             $$PWD/rssdownloadrule.cpp \
-             $$PWD/rssdownloadrulelist.cpp \
              $$PWD/cookiesdlg.cpp \
-             $$PWD/rssfile.cpp \
-             $$PWD/rssparser.cpp \
              $$PWD/htmlbrowser.cpp
 
 FORMS +=   $$PWD/rss.ui \
